remove the copying of patch pysical size

+++ src/data_structs/MPR_local_patch_structs.h
@@ -16,7 +16,7 @@ struct mpr_local_patch_struct
 
 	unsigned char* buffer;                   /* data buffer */
 
-	unsigned long long out_file_size;        /* the size out file */
+	int out_file_size;        /* the size out file */
 
 	MPR_patch *patch;                        /* Pointer to the patches that are included in the local dataset */
